feat: start host-metering service on rpm installation
The host-metering.service is started at the end of transaction if the
service is enabled. It is enabled by preset, so it is possible to
disable it also by other preset with higher priority.

It's in selinux poststrans as it was otherwise failing with permissions
issues.

Overall this is a hack and e.g. might be against Fedora packaging
policy.

https://issues.redhat.com/browse/HMS-3358

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>